requirements.txt
# Shared
PyQt6>=6.4.0
Pillow>=10.0.0
numpy>=1.24.0

# PySnagit - Screen capture
mss>=9.0.0
opencv-python>=4.8.0
pyaudio>=0.2.13
keyboard>=0.13.5
pywin32>=306
imageio>=2.31.0
imageio-ffmpeg>=0.4.8

# PyPDF Editor - PDF editing
PyMuPDF>=1.23.0
pytesseract>=0.3.10
